// src/types/ShoppingType.d.ts
export type ShoppingDetailInfoType = {
  contentid: string;
  contenttypeid: string;
  serialnum: string;
  infoname: string;
  infotext: string;
  fldgubun: string;
};

export type ShoppingDetailIntroType = {
  contentid: string;
  contenttypeid: string;
  saleitem: string;
  saleitemcost: string;
  fairday: string;
  opendateshopping: string;
  shopguide: string;
  culturecenter: string;
  restroom: string;
  infocentershopping: string;
  scaleshopping: string;
  restdateshopping: string;
  parkingshopping: string;
  chkbabycarriageshopping: string;
  chkpetshopping: string;
  chkcreditcardshopping: string;
  opentime: string;
};